Move IsPath protocol into its own file
Path module is now a clean file. There's still some more cleanup, but
things are coming along nicely.
